reaper_controller/requirements.txt
# Copyright © 2024 Lunarbase (https://lunarbase.ai/) <contact@lunarbase.ai>
#
# SPDX-License-Identifier: GPL-3.0-or-later

langchain~=0.1.7
python-reapy~=0.10.0